+++ instagram/src/App.js
@@ -1,26 +1,24 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
+import SearchBar from '../src/Components/SearchBar';
+
+
 
 class App extends Component {
+  constructor() {
+    super();
+    this.state = {
+      dummyObj: []
+    };
+  }
+
+
   render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <p>
-            Edit <code>src/App.js</code> and save to reload.
-          </p>
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-        </header>
+    return(
+      <div className='app-wrap'>
+        <SearchBar />
       </div>
+      
+      
     );
   }
 }
